moving TrustedDealerKeyPackages to keypackages new file
