+++ ironfish-rust-nodejs/src/lib.rs
@@ -2,7 +2,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-use std::collections::HashMap;
 use std::fmt::Display;
 
 use ironfish::keys::Language;
@@ -169,14 +168,3 @@ pub fn is_valid_public_address(hex_address: String) -> bool {
     PublicAddress::from_hex(&hex_address).is_ok()
 }
 
-#[napi(object)]
-pub struct TrustedDealerKeyPackages {
-    pub verifying_key: String,
-    pub proof_generation_key: String,
-    pub view_key: String,
-    pub incoming_view_key: String,
-    pub outgoing_view_key: String,
-    pub public_address: String,
-    pub key_packages: HashMap<String, String>,
-    pub public_key_package: String,
-}
