+++ ironfish-rust-nodejs/src/structs/mod.rs
@@ -19,3 +19,6 @@ pub use transaction::*;
 
 mod witness;
 pub use witness::*;
+
+mod key_packages;
+pub use key_packages::*;
